Log management: performance improvements
Do not systematically save the log metadata when inserting a new
datapoint, but let the app trigger it when necessary. In particular,
this ensures that when saving a device log, we do not update the metadata
as many times as we have data points, but only once the datapoints are
all saved.